tweak
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@12679 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/switch_core_port_allocator.c
@@ -118,7 +118,7 @@ SWITCH_DECLARE(switch_status_t) switch_core_port_allocator_request_port(switch_c
 	int odd = switch_test_flag(alloc, SPF_ODD);
 
 	switch_mutex_lock(alloc->mutex);
-	srand((unsigned)(intptr_t)port_ptr + (uint32_t)switch_thread_self() + switch_micro_time_now());
+	srand((unsigned)(intptr_t)port_ptr + (unsigned)(intptr_t)switch_thread_self() + switch_micro_time_now());
 
 	while (alloc->track_used < alloc->track_len) {
 		uint32_t index;
